% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataCheck.R
\name{dataCheck}
\alias{dataCheck}
\title{dataCheck.R}
\usage{
dataCheck(demo.data = demo_data, psa.data = psa_data, bx.data = bx_data,
  surg.data = surg_data)
}
\arguments{
\item{demo.data}{Patient-level data. This dataframe should contain one record (row) per patient. Variables (columns) must include a unique identifier (ID) and valid date of birth for each patient.}

\item{psa.data}{PSA data. This dataframe should contain one record per PSA test (per patient). Variables must include patient ID, date of test, and total PSA observed.}

\item{bx.data}{Biopsy data. This dataframe should contain one record per biopsy (per patient). Variables must include patient ID, date of biopsy, indicator of grade reclassication, and prostate volume (if assessed).}

\item{surg.data}{Surgery data. This dataframe should contain one record per surgery performed. Variables must include patient ID, surgery date, and Gleason score (GS) assessed on entire specimen.}
}
\description{
Check if loaded patient data has values within acceptable range. See vignette for restrictions on data.
}

