% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacheck2.R
\name{dataCheck2}
\alias{dataCheck2}
\title{dataCheck.R}
\usage{
dataCheck2(tx.data, demo.data, psa.data, bx.data)
}
\arguments{
\item{tx.data}{One record per treatment received per patient containing treatment date and GS}

\item{demo.data}{Demographic data, one record per patient ID containing DOB}

\item{psa.data}{PSA data, one record per PSA test per patient containing date of PSA test}

\item{bx.data}{One record per biopsy per patient containing reclassicfication, volume, and dx}
}
\description{
Check if loaded patient data has values within acceptable range.
}

