% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{distance}
\alias{distance}
\title{distance}
\usage{
distance(pt1 = 5, pt2 = 38)
}
\arguments{
\item{idInput}{Integer of patient record to be displayed}
}
\description{
Find the distance between two patients
}

