% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillPatientTablesIOP.R
\name{fillPatientTablesIOP}
\alias{fillPatientTablesIOP}
\title{fillPatientTablesIOP.R}
\usage{
fillPatientTablesIOP(demo.data = demo_data, psa.data = psa_data,
  bx.data = bx_data, surg.data = surg_data)
}
\arguments{
\item{demo.data}{Demographic data, one record per patient ID containing DOB}

\item{psa.data}{PSA data, one record per PSA test per patient containing date of PSA test}

\item{bx.data}{One record per biopsy per patient containing reclassicfication, volume, and dx}

\item{surg.data}{One record per treatment received per patient containing treatment date and GS}
}
\value{
List of three dataframes, pt.data, psa.data, bx.full
}
\description{
Load dataframes from four patient data files to create a list of formatted patient dataframes.
}

