% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIndividualData.R
\name{getIndividualData}
\alias{getIndividualData}
\title{getIndividualData.R}
\usage{
getIndividualData(idInput = 5, patientDataframes, bx_data)
}
\arguments{
\item{idInput}{Integer of patient record to be displayed}

\item{patientDataframes}{Full}
}
\description{
tet a specified patient's demographic, PSA, biopsy, and treatment results to the console
}

