% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIndividualData.R
\name{plotIndividualData}
\alias{plotIndividualData}
\title{plotIndividualData}
\usage{
plotIndividualData(pt.id, what.data = "both", log.scale = T,
  plot.psad = F)
}
\arguments{
\item{pt.id}{Patient whose data to print}

\item{what.data}{What clinical data source to print}

\item{log.scale}{T}

\item{plot.psad}{F}
}
\description{
Find the distance between two patients.
}

