% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPtData.R
\name{plotPtData}
\alias{plotPtData}
\title{Plot patient data using the "shaped" data. This function isn't necessary to prep the data and arguments for jags.}
\usage{
plotPtData(pt.id, what.data = "both", log.scale = T, plot.psad = F)
}
\arguments{
\item{pt.id}{Integer of patient record to be displayed}

\item{what.data}{Variable that says what clinical data source to plot. "Both" = plot both PSA and Biopsy data, "psa" = plot just PSA data, "biopsy" = plot just biopsy data}

\item{log.scale}{TRUE if using log scale for PSA, FALSE if not. TRUE default.}

\item{plot.psad}{TRUE if using PSA Density instead of PSA, FALSE if not. FALSE defaul}
}
\description{
Plot patient data using the "shaped" data. This function isn't necessary to prep the data and arguments for jags.
}

