% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printIndividualData.R
\name{printIndividualData}
\alias{printIndividualData}
\title{printIndividualData.R}
\usage{
printIndividualData(idInput = 10, patientDataframes)
}
\arguments{
\item{idInput}{Integer of patient record to be displayed}

\item{patientDataframes}{Full}
}
\description{
Print a specified patient's demographic, PSA, biopsy, and treatment results to the console
}

