% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RJAGSprep.R
\name{RJAGSprep}
\alias{RJAGSprep}
\title{RJAGSprep.R}
\usage{
RJAGSprep(patientDataframes = patientDataframes, modelFile = model.file)
}
\arguments{
\item{pt.data}{Filled dataframe of one record per patient containing DOB, age at dx, vol, and rc}

\item{psa.data}{Filled dataframe of one record per patient PSA reading containing patient subj ID, date of PSA, log PSA, and vol}

\item{bx.data}{Filled dataframe of one record per patient biopsy, containing patient subj ID, time, rc, and age}
}
\value{
List of parameters for RJAGS prep


WORKFLOW
0. Load packages and  necessary data.
1. Format pt-level data for JAGS run
2. Format PSA data for JAGS run
3. Format biopsy data for JAGS run
4. Load libraries
5. Define data to be sent to jags function
6. Initialize model parameters
7. Define parameters to be tracked
8. Define other jags settings
9. Write model definition
}
\description{
Prep files for RJAGS by creating list of parameters for RJAGS run.
}

