% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeJAGSmodel.R
\name{writeJAGSmodel}
\alias{writeJAGSmodel}
\title{Write JAGS model from text file}
\usage{
writeJAGSmodel(file.name = "UNADJ-jags-model.txt")
}
\arguments{
\item{file.name}{the name of the file that describes the model}
}
\value{
the model
}
\description{
Write JAGS model from text file
}

