% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeJAGSmodelOP.R
\name{writeJAGSmodelIOP}
\alias{writeJAGSmodelIOP}
\title{Write JAGS model from IOP text file}
\usage{
writeJAGSmodelIOP(file.name = "IOP-jags-model.txt")
}
\arguments{
\item{file.name}{the name of the file that describes the model}
}
\value{
the model
}
\description{
Write JAGS model from IOP text file
}

