% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeModel.R
\name{writeModel}
\alias{writeModel}
\title{writeModel}
\usage{
writeModel(file.name = "UNADJ-jags-model.txt")
}
\arguments{
\item{The}{name of the model file to create}
}
\description{
Check if loaded patient data has values within acceptable range.
}

