% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closestK.R
\name{closestK}
\alias{closestK}
\title{closestK}
\usage{
closestK(ptId = 6, K = 25, D = 1, patientDataframes)
}
\arguments{
\item{K}{Number of nearby patients to return}

\item{D}{Max distance between comparable patients}

\item{pt.id}{Integer of patient record to be displayed}
}
\description{
Find the distance between two patients
}

