% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closestK.R
\name{closestK}
\alias{closestK}
\title{closestK}
\usage{
closestK(pt.id = 5, K = 50, D = 1)
}
\arguments{
\item{pt.id}{Integer of patient record to be displayed}

\item{K}{Number of nearby patients to return}

\item{D}{Max distance between comparable patients}
}
\description{
Find the distance between two patients
}

