% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataCheck.R
\name{dataCheck}
\alias{dataCheck}
\title{dataCheck.R}
\usage{
dataCheck(surg.data = surg_data, demo.data = demo_data,
  psa.data = psa_data, bx.data = bx_data)
}
\arguments{
\item{surg.data}{One record per treatment received per patient containing treatment date and GS}

\item{demo.data}{Demographic data, one record per patient ID containing DOB}

\item{psa.data}{PSA data, one record per PSA test per patient containing date of PSA test}

\item{bx.data}{One record per biopsy per patient containing reclassicfication, volume, and dx}
}
\description{
Check if loaded patient data has values within acceptable range.
}

