% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataCheck.R
\name{dataCheck}
\alias{dataCheck}
\title{Check if data has values within specified range}
\usage{
dataCheck(tx.data, demo.data, psa.data, bx.data)
}
\arguments{
\item{tx.data}{one record per treatment received per patient}

\item{demo.data}{demographic data, one record per patient}

\item{psa.data}{psa data, one record per PSA test per patient}

\item{bx.data}{one record per biopsy per patient}
}
\value{
true if error was caught, false if not
}
\description{
Check if data has values within specified range
}

