% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{distance}
\alias{distance}
\title{distance}
\usage{
distance(ptId, comp, distDataframe, stdev)
}
\arguments{
\item{ptId}{Patient 1}

\item{comp}{Patient 2, to be compared to}

\item{distDataframe}{Dataframe of variables to be compared in distance function}
}
\description{
Find the distance between two patients
}

