% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillPatientTables.R
\name{fillPatientTables}
\alias{fillPatientTables}
\title{Load dataframes from four patient data files}
\usage{
fillPatientTables(tx.data = tx_data, demo.data = demo_data,
  psa.data = psa_data, bx.data = bx_data)
}
\arguments{
\item{tx.data}{one record per treatment received per patient}

\item{demo.data}{demographic data, one record per patient}

\item{psa.data}{psa data, one record per PSA test per patient}

\item{bx.data}{one record per biopsy per patient}
}
\value{
list of three dataframes, pt.data, psa.data, bx.full
}
\description{
Load dataframes from four patient data files
}

