% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RJAGSprep.R
\name{RJAGSprep}
\alias{RJAGSprep}
\title{RJAGSprep.R}
\usage{
RJAGSprep(pt, IOP = TRUE)
}
\arguments{
\item{pt}{Dataframe of formatted active surveillance data created by the function fillPatientTables()}

\item{IOP}{TRUE if biopsy and surgery occurences are informative of the underlying GS, FALSE if non-informative.}
}
\value{
List of arguments for running analysis using RJAGS: \code{jags_data} (data), \code{inits} (inits), \code{parameters.to.save} (parameters.to.save).
}
\description{
Prepare necessary arguments for running prostate cancer later class analysis using RJAGS.
}

