% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RJAGSprep.R
\name{RJAGSprep}
\alias{RJAGSprep}
\title{Prep files for RJAGS run}
\usage{
RJAGSprep(patientDataframes = patientDataframes, model.file)
}
\arguments{
\item{pt.data}{filled dataframe}

\item{psa.data}{filled dataframe}

\item{bx.data}{filled dataframe}
}
\value{
list for RJAGS prep
}
\description{
Prep files for RJAGS run
}

